+++ src/main/java/uk/gov/companieshouse/filing/reader/FilingReaderImpl.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.reader;
+package uk.gov.companieshouse.filing.reader;
 
 import java.util.ArrayList;
 import java.util.Arrays;
