+++ src/main/java/uk/gov/companieshouse/filing/writer/FilingWriterException.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.writer;
+package uk.gov.companieshouse.filing.writer;
 
 public class FilingWriterException extends Exception {
